docs: UserStatus 서비스 생성
